core: Add rules config and improve mode value parse.
- 0, off -> SeverityMode::Off
- 1, error -> SeverityMode::Error
- 2, warning -> SeverityMode::Warning
