+++ autocorrect/.autocorrectrc.default
@@ -1,5 +1,11 @@
+rules:
+  space-word: 1
+  space-punctuation: 1
+  fullwidth: 1
+  halfwidth: 1
+  no-space-fullwidth: 1
 spellcheck:
-  mode: 0
+  mode: 2
   words:
     # Please do not add a general English word (eg. apple, python) here.
     # Users can add their special words to their .autocorrectrc file by their need.
