+++ autocorrect/src/config/mod.rs
@@ -1,0 +1,7 @@
+mod config;
+mod severity;
+mod spellcheck;
+
+pub use config::*;
+pub use severity::*;
+pub use spellcheck::*;
